[Jihwan] Add Coin Value Count Up, Add Player Aim Animation, Fix Player Run And Shoot Animation

+++ KillingFloor/Assets/00_KillingFloor/Scripts/Player/LivingEntity.cs
@@ -8,7 +8,7 @@ public class LivingEntity : MonoBehaviour, IDamageable
     public float startingHealth = 100f; // 시작 체력
     public float health { get; protected set; } // 현재 체력
     public float armor { get; protected set; }
-    public float coin { get; protected set; }
+    public int coin { get; protected set; }
     public bool dead { get; protected set; } // 사망 상태
     public event Action onDeath; // 사망시 발동할 이벤트
